services/observability-service/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1
boto3==1.37.18
psycopg2-binary==2.9.10
timescaledb==0.0.2
structlog==25.1.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
